requirements.txt
lxml==4.1.1
requests==2.18.4
celery==4.1.0
neo4j-driver==1.5.1
fake-useragent==0.1.10
dill==0.2.7.1
numpy==1.13.3